agent-driven: move stuff into src/format
