+++ components/datetime/src/provider/fields/mod.rs
@@ -7,12 +7,10 @@
 #![allow(clippy::exhaustive_structs)] // Field and FieldULE part of data struct
 
 mod length;
-mod numeric_override;
 pub(crate) mod symbols;
 
 use displaydoc::Display;
-pub use length::{FieldLength, LengthError};
-pub use numeric_override::FieldNumericOverrides;
+pub use length::{FieldLength, FieldNumericOverrides, LengthError};
 pub use symbols::*;
 use writeable::Writeable;
